+++ HomeScreen/TwitterRESTAPI.swift
@@ -42,24 +42,27 @@ func Search(completion: (result: [String]) -> Void) {
                     (response, data, connectionError) -> Void in
                     if (connectionError == nil) {
 
-                        // Setup a tweet array to contain all of those juicy tweets
                         var tweetArray = Array<String>()
 
-                        var jsonError : NSError?
-                        let json : AnyObject? =
-                        NSJSONSerialization.JSONObjectWithData(data,
-                            options: nil,
-                            error: &jsonError)
+                        if let responseData = data {
+                            var jsonError : NSError?
+                            let json : AnyObject? =
+                            NSJSONSerialization.JSONObjectWithData(responseData,
+                                options: nil,
+                                error: &jsonError)
 
-                        // Iterate through JSON response and append the values to the TweetArray
-                        if let statuses = json!["statuses"] as? JSONArray {
+                            // Iterate through JSON response and append the values to the TweetArray
+                            if let jsonDictionary = json as? JSONDictionary {
+                                if let statuses = jsonDictionary["statuses"] as? JSONArray {
 
-                            // For each tweet in the status block of the json request e.g. {"statuses": [tweets.........
-                            for tweet in statuses {
-                                if let id = tweet["id_str"] as?String{
+                                    // For each tweet in the status block of the json request e.g. {"statuses": [tweets.........
+                                    for tweet in statuses {
+                                        if let id = tweet["id_str"] as?String{
 
-                                    // Append the Tweet to the array
-                                    tweetArray.append(id)
+                                            // Append the Tweet to the array
+                                            tweetArray.append(id)
+                                        }
+                                    }
                                 }
                             }
                         }
@@ -85,6 +88,5 @@ func Search(completion: (result: [String]) -> Void) {
         
     }
 }
-